requirements.txt
absl-py==1.4.0
ai-safety-gridworlds @ https://github.com/timorl/ai-safety-gridworlds/tarball/master
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.7
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
Deprecated==1.2.13
distrax==0.1.3
dm-tree==0.1.8
equinox==0.10.3
exceptiongroup==1.1.1
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.27.1
gymnasium-notices==0.0.1
h5py==3.8.0
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
iniconfig==2.0.0
jax==0.4.6
jax-jumpy==0.2.0
jaxlib==0.4.6
jaxtyping==0.2.14
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
moviepy==1.0.3
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.22.4
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycolab==1.2
pycparser==2.21
pygame==2.1.3.dev8
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safe-grid-gym @ https://github.com/david-lindner/safe-grid-gym/tarball/master
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==2.2.0
thefuzz==0.19.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typeguard==3.0.2
types-tabulate==0.9.0.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
